+++ workers/ohsome_quality_analyst/ohsome/client.py
@@ -1,6 +1,7 @@
 import datetime
+import json
 import logging
-from typing import Dict
+from typing import Dict, Optional
 
 import httpx
 
@@ -9,27 +10,42 @@ from ohsome_quality_analyst.utils.definitions import OHSOME_API
 
 # TODO: Add documentation on time string format.
 # TODO: Add tests for ohsome package, including time = None leading to errors
-async def query(layer, bpolys: str, time: str = None, endpoint: str = None) -> Dict:
+async def query(
+    layer,
+    bpolys: str,
+    time: Optional[str] = None,
+    endpoint: Optional[str] = None,
+    ratio: bool = False,
+) -> Dict:
     """Query ohsome API endpoint with filter."""
-    if endpoint:
+    if endpoint is not None:
         url = OHSOME_API + endpoint
     else:
         url = OHSOME_API + layer.endpoint
-    data = {"bpolys": bpolys, "filter": layer.filter}
+
+    if ratio:
+        url = url + "/ratio"
+        data = {"bpolys": bpolys, "filter": layer.filter, "filter2": layer.ratio_filter}
+    else:
+        data = {"bpolys": bpolys, "filter": layer.filter}
+
     if time is not None:
         data["time"] = time
+
+    logging.info("Query ohsome API.")
+    logging.debug("Query data: " + json.dumps(data, indent=4, sort_keys=True))
+
     # set custom timeout as ohsome API takes a long time to send an answer
     timeout = httpx.Timeout(5, read=600)
     async with httpx.AsyncClient(timeout=timeout) as client:
         resp = await client.post(url, data=data)
+    logging.debug(
+        "Query response: " + json.dumps(resp.json(), indent=4, sort_keys=True)
+    )
 
-    logging.info("Query ohsome API.")
-    logging.debug("Query URL: " + url)
-    logging.debug("Query Filter: " + layer.filter)
     if resp.status_code == 200:
         logging.info("Query successful!")
     elif resp.status_code == 404:
-        # TODO: Handle when query fails
         logging.info("Query failed!")
 
     return resp.json()