+++ workers/tests/integrationtests/test_api_report.py
@@ -11,7 +11,7 @@ import unittest
 
 import geojson
 from fastapi.testclient import TestClient
-from schema import Schema
+from schema import Optional, Or, Schema
 
 from ohsome_quality_analyst.api import app
 
@@ -56,6 +56,7 @@ class TestApiReport(unittest.TestCase):
                             "description": str,
                             "endpoint": str,
                             "filter": str,
+                            Optional("ratio_filter", default=None): Or(str, None),
                         },
                         "result": {
                             "value": float,